% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smsBatch.R
\name{smsBatch}
\alias{smsBatch}
\title{smsBatch}
\usage{
smsBatch(files = "all", removeDuplicates = TRUE, keepUnsent = FALSE,
  charCount = TRUE, noMMS = FALSE)
}
\arguments{
\item{files}{A character vector of files to import. Default "all" will
import all XML files from the working directory.}

\item{removeDuplicates}{When TRUE (default), will remove duplicate rows;
useful for combining overlapping backups.}

\item{keepUnsent}{Keeps unsent Draft and Failed messages when TRUE.
Defaults to FALSE}

\item{charCount}{When TRUE, adds a column counting the characters in each
text.}

\item{noMMS}{A flag to specify when no MMS data is present. If no MMS data
exists, MMS munging code will throw errors. When TRUE, MMS munging code is
skipped.}
}
\value{
A data.frame of text message data, sorted by date.
}
\description{
Convenience wrapper for smsProcess, to import and combine mulitple files
}
\details{
smsBatch is a convenience wrapper for smsProcess, which pulls
and combines multiple XML files of text message records. It can be passed
a character string of file names to pass to process, or by default it will
pull all XML files in the working directory.
}
\examples{
#Import all XML files in working directory with default options
smsProcess()
#Import specific files, passing options to smsProceess
smsProcess(c("backup1.xml", "backup2.xml"), keepUnsent = TRUE,
charCount = FALSE, noMMS = TRUE)
}
\seealso{
\code{\link{smsProcess}} For the documentation on the underlying
function called by smsBatch
}
\keyword{android}
\keyword{messages,}
\keyword{mms,}
\keyword{sms,}
\keyword{text}

